+++ client/components/Admin/Template/TemplateDetail/TemplateDetail.tsx
@@ -3,8 +3,8 @@ import { BiTrash } from 'react-icons/bi';
 import styled from 'styled-components';
 import { styleMixin } from '@/src/shared/styles';
 import { keys } from '@/src/shared/utils';
-import useTemplateManager from '../../hooks/useTemplateManager';
-import useTemplateStoreManager from '../../hooks/useTemplateStoreManager';
+import useTemplateManager from '../hooks/useTemplateManager';
+import useTemplateStoreManager from '../hooks/useTemplateStoreManager';
 
 function NoTemplateSelected() {
   return <WarringMsg>현재 선택된 리스트가 없습니다.</WarringMsg>;
