+++ client/pages/coverLetter/index.tsx
@@ -1,30 +1,1 @@
-import { useEffect } from 'react';
-
-import styled from 'styled-components';
-import { styleMixin } from '@/src/shared/styles';
-
-import CoverLetter from '@/components/CoverLetter/CoverLetter';
-import { useCoverLetterStore } from '@/src/entities/coverLetter';
-
-function CoverLetterPage() {
-  const { initCoverLetter } = useCoverLetterStore();
-
-  useEffect(
-    () => () => {
-      initCoverLetter();
-    },
-    [],
-  );
-
-  return (
-    <CoverLetterStyle>
-      <CoverLetter />
-    </CoverLetterStyle>
-  );
-}
-
-export default CoverLetterPage;
-
-const CoverLetterStyle = styled.div`
-  ${styleMixin.Column()}
-`;
+export { default } from '@/src/pages/coverletter';
