+++ client/src/features/board/index.ts
@@ -1,0 +1,2 @@
+export { default as BoardList } from './BoardList';
+export { default as BoardListHead } from './BoardListHead';
